Gracefully handle case where JsonWebToken is injected and JwtAuth provider is not in use.
